+++ src/main/java/com/lapissea/j2/pong/game/Message.java
@@ -2,5 +2,5 @@ package com.lapissea.j2.pong.game;
 
 import java.io.Serializable;
 
-public record Message(int profileId, String text) implements Serializable{
+public record Message(long profileId, String text) implements Serializable{
 }
